mapper, controller, service update

+++ backendQLKS/src/main/java/com/example/backendqlks/dao/UserRolePermissionRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface UserRolePermissionRepository extends JpaRepository<UserRolePermission, Long> {
+public interface UserRolePermissionRepository extends JpaRepository<UserRolePermission, Integer> {
     List<UserRolePermission> findByUserRoleId(int id);
 
     List<UserRolePermission> findByPermissionId(int id);
