+++ backendQLKS/src/main/java/com/example/backendqlks/dao/UserRoleRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface UserRoleRepository extends JpaRepository<UserRole, Long> {
+public interface UserRoleRepository extends JpaRepository<UserRole, Integer> {
     List<UserRole> findByNameIgnoreCase(String name);
 }
